Remove workload endpoint management
