+++ calicoctl/commands/resources.go
@@ -30,29 +30,6 @@ import (
 	"github.com/tigera/libcalico-go/lib/net"
 )
 
-// Create a new CalicoClient using connection information in the specified
-// filename (if it exists), dropping back to environment variables for any
-// parameter not loaded from file.
-func newClient(cf string) (*client.Client, error) {
-	if _, err := os.Stat(cf); err != nil {
-		glog.V(2).Infof("Config file cannot be read - reading config from environment")
-		cf = ""
-	}
-
-	cfg, err := client.LoadClientConfig(cf)
-	if err != nil {
-		return nil, err
-	}
-	glog.V(2).Infof("Loaded client config: type=%v %#v", cfg.BackendType, cfg.BackendConfig)
-
-	c, err := client.New(*cfg)
-	if err != nil {
-		return nil, err
-	}
-
-	return c, err
-}
-
 // Convert loaded resources to a slice of resources for easier processing.
 // The loaded resources may be a slice containing resources and resource lists, or
 // may be a single resource or a single resource list.  This function handles the
@@ -182,7 +159,7 @@ type commandResults struct {
 	resources []unversioned.Resource
 }
 
-// Common function for configuration commands create, replace and delete.  All
+// Common function for configuration commands apply, create, replace and delete.  All
 // these commands:
 // 	-  Load resources from file (or if not specified determine the resource from
 // 	   the command line options).
